+++ extensions/tiktok/src/main/java/app/morphe/extension/tiktok/settings/SettingsStatus.java
@@ -45,6 +45,7 @@ public class SettingsStatus {
     public static boolean disableLongPressQuickShareEnabled = false;
     public static boolean disableLongPressRepostEnabled = false;
     public static boolean nonPersonalizedSearchEnabled = false;
+    public static boolean hideSearchSuggestionsEnabled = false;
     public static boolean liveSearchEnabled = false;
     public static boolean seekbarThumbnailEnabled = false;
     public static boolean stopVideoLoopingEnabled = false;
@@ -126,6 +127,10 @@ public class SettingsStatus {
 
     public static void enableNonPersonalizedSearch() {
         nonPersonalizedSearchEnabled = true;
+    }
+
+    public static void enableHideSearchSuggestions() {
+        hideSearchSuggestionsEnabled = true;
     }
 
     public static void enableLiveSearch() {